# swagger.yaml
openapi: 3.0.0
info:
  title: Runway Kicks API
  version: 1.0.0
  description: API documentation for the Runway Kicks backend
servers:
  - url: http://localhost:7000
    description: Local development server
paths:
...
  /men/filter:
    get:
      summary: Filter men's products by category and/or color
      tags:
        - Men
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          example: "Clothing"
        - name: color
          in: query
          required: false
          schema:
            type: string
          example: "Red"
      responses:
        "200":
          description: Filtered list of men's products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Men's T-Shirt"
                    price: 25.99
                    category: "Clothing"
                    color: "Red"
                    offer: false
...
  /men/sort:
    get:
      summary: Sort men's products by price or title
      tags:
        - Men
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - lth
              - htl
              - asc
              - desc
          example: "lth"
      responses:
        "200":
          description: Sorted list of men's products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Men's T-Shirt"
                    price: 19.99
                    category: "Clothing"
                    color: "Blue"
                    offer: false
  /men/search:
    get:
      summary: Search for men's products by title
      tags:
        - Men
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          example: "Shirt"
      responses:
        "200":
          description: List of men's products matching the search term
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Men's Shirt"
                    price: 29.99
                    category: "Clothing"
                    color: "White"
                    offer: false
  /men/create:
    post:
      summary: Create a new men's product (admin only)
      tags:
        - Men
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - price
                - category
                - color
                - offer
              properties:
                title:
                  type: string
                  example: "Men's Jacket"
                price:
                  type: number
                  format: float
                  example: 79.99
                category:
                  type: string
                  example: "Clothing"
                color:
                  type: string
                  example: "Black"
                offer:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "New product created"
  /men/update/{id}:
    patch:
      summary: Update a men's product (admin only)
      tags:
        - Men
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Men's Jacket"
                price:
                  type: number
                  format: float
                  example: 89.99
                category:
                  type: string
                  example: "Clothing"
                color:
                  type: string
                  example: "Red"
                offer:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product updated"
  /men/delete/{id}:
    delete:
      summary: Delete a men's product (admin only)
      tags:
        - Men
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product deleted"
...
  /women/filter:
    get:
      summary: Filter women's products by category and/or color
      tags:
        - Women
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          example: "Clothing"
        - name: color
          in: query
          required: false
          schema:
            type: string
          example: "Red"
      responses:
        "200":
          description: Filtered list of women's products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Women's Dress"
                    price: 39.99
                    category: "Clothing"
                    color: "Red"
                    offer: false
...
  /women/sort:
    get:
      summary: Sort women's products by price or title
      tags:
        - Women
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - lth
              - htl
              - asc
              - desc
          example: "lth"
      responses:
        "200":
          description: Sorted list of women's products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Women's Dress"
                    price: 29.99
                    category: "Clothing"
                    color: "Blue"
                    offer: false
  /women/search:
    get:
      summary: Search for women's products by title
      tags:
        - Women
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          example: "Dress"
      responses:
        "200":
          description: List of women's products matching the search term
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Women's Dress"
                    price: 29.99
                    category: "Clothing"
                    color: "Blue"
                    offer: false
  /women/create:
    post:
      summary: Create a new women's product (admin only)
      tags:
        - Women
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - price
                - category
                - color
                - offer
              properties:
                title:
                  type: string
                  example: "Women's Jacket"
                price:
                  type: number
                  format: float
                  example: 79.99
                category:
                  type: string
                  example: "Clothing"
                color:
                  type: string
                  example: "Black"
                offer:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "New product created"
  /women/update/{id}:
    patch:
      summary: Update a women's product (admin only)
      tags:
        - Women
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Women's Jacket"
                price:
                  type: number
                  format: float
                  example: 89.99
                category:
                  type: string
                  example: "Clothing"
                color:
                  type: string
                  example: "Red"
                offer:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product updated"
  /women/delete/{id}:
    delete:
      summary: Delete a women's product (admin only)
      tags:
        - Women
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product deleted"
...
  /order:
    get:
      summary: Get all orders for the authenticated user
      tags:
        - Order
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of orders for the authenticated user
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    orderID:
                      type: string
                    userID:
                      type: string
                    productID:
                      type: string
                    quantity:
                      type: integer
                    totalPrice:
                      type: number
                      format: float
                    status:
                      type: string
                      enum: [pending, completed, cancelled]
                example:
                  - orderID: "60c72b2f9f1b2c3d8c1e1d1a"
                    userID: "60c72b2f9f1b2c3d8c1e1d1b"
                    productID: "60c72b2f9f1b2c3d8c1e1d2a"
                    quantity: 2
                    totalPrice: 39.98
                    status: "completed"
                  - orderID: "60c72b2f9f1b2c3d8c1e1d2b"
                    userID: "60c72b2f9f1b2c3d8c1e1d1b"
                    productID: "60c72b2f9f1b2c3d8c1e1d2c"
                    quantity: 1
                    totalPrice: 19.99
                    status: "pending"
  /order/search:
    get:
      summary: Search orders by product title for the authenticated user
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          example: "Nike"
      responses:
        "200":
          description: List of orders that match the search query
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    orderID:
                      type: string
                    userID:
                      type: string
                    productID:
                      type: string
                    quantity:
                      type: integer
                    totalPrice:
                      type: number
                      format: float
                    status:
                      type: string
                      enum: [pending, completed, cancelled]
                example:
                  - orderID: "60c72b2f9f1b2c3d8c1e1d1a"
                    userID: "60c72b2f9f1b2c3d8c1e1d1b"
                    productID: "60c72b2f9f1b2c3d8c1e1d2a"
                    quantity: 2
                    totalPrice: 39.98
                    status: "completed"
        "400":
          description: No orders found for the specified title
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "No orders found for the specified title"